Giveup UI 추가

+++ Assets/Scripts/DataTable/PlayerData.cs
@@ -5,8 +5,8 @@ public class PlayerData
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public int SkillCoolDown { get; set; }
+    public float SkillCoolDown { get; set; }
     public int HealthPoint { get; set; }
-    public int MoveSpeed { get; set; }
+    public float MoveSpeed { get; set; }
     public string KeyIds { get; set; }
 }
